+++ everscale/contracts/bridge/interfaces/multivault/IMultiVaultSolanaEverscaleEventAlien.sol
@@ -20,7 +20,7 @@ interface IMultiVaultSolanaEverscaleEventAlien {
         string name_,
         string symbol_,
         uint8 decimals_,
-        uint128 amount_,
+        uint64 amount_,
         address recipient_,
         address proxy_,
         address token_
